pyproject.toml
[project]
name = "funnel-canary"
version = "0.0.2"
description = "AI Agent for closed-loop problem solving"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
